+++ config/smsapi.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'token' => 'tGnIPugtU21BgPfgEQe5EZwnGCIqyM8urmUbZ11X',
+    'from' => 'Info'
+];