+++ resources/lang/pl/item_status.php
@@ -4,5 +4,5 @@
         'not_ordered' => 'Niezamówione',
         'delivered' => 'Dostarczone',
         'unknown' => 'Nieznany',
-        'partailly_delivered' => 'Częściowo zrealizowany',
+        'partially_delivered' => 'Częściowo zrealizowany',
     ];