+++ resources/views/administrator/smsapi.blade.php
@@ -1,0 +1,11 @@
+@extends('auth.auth_layout')
+
+@section('content')
+    <div class="row">
+        <div class="col">
+            {{ $smsResult }}
+        </div>
+    </div>
+@endsection
+
+